move Piece and Capture to types.py
