+++ stockfish/models.py
@@ -11,10 +11,11 @@ from typing import Any, List, Optional, Union, Dict, Tuple
 import copy
 import os
 from dataclasses import dataclass
-from enum import Enum
 import re
 import datetime
 import warnings
+
+from stockfish.types import Piece, Capture
 
 
 class Stockfish:
@@ -1028,7 +1029,7 @@ class Stockfish:
             )
         rank_visual: str = self.get_board_visual().splitlines()[17 - 2 * rank_num]
         piece_as_char: str = rank_visual[2 + (ord(file_letter) - ord("a")) * 4]
-        return None if piece_as_char == " " else Stockfish.Piece(piece_as_char)
+        return None if piece_as_char == " " else Piece(piece_as_char)
 
     def will_move_be_a_capture(self, move_value: str) -> Capture:
         """Returns whether the proposed move will be a direct capture,
@@ -1050,34 +1051,34 @@ class Stockfish:
         """
         if not self.is_move_correct(move_value):
             raise ValueError("The proposed move is not valid in the current position.")
-        starting_square_piece: Optional[Stockfish.Piece] = self.get_what_is_on_square(
+        starting_square_piece: Optional[Piece] = self.get_what_is_on_square(
             move_value[:2]
         )
-        ending_square_piece: Optional[Stockfish.Piece] = self.get_what_is_on_square(
+        ending_square_piece: Optional[Piece] = self.get_what_is_on_square(
             move_value[2:4]
         )
         if ending_square_piece is not None:
             if not self._parameters["UCI_Chess960"]:
-                return Stockfish.Capture.DIRECT_CAPTURE
+                return Capture.DIRECT_CAPTURE
             else:
                 # Check for Chess960 castling:
                 castling_pieces = [
-                    [Stockfish.Piece.WHITE_KING, Stockfish.Piece.WHITE_ROOK],
-                    [Stockfish.Piece.BLACK_KING, Stockfish.Piece.BLACK_ROOK],
+                    [Piece.WHITE_KING, Piece.WHITE_ROOK],
+                    [Piece.BLACK_KING, Piece.BLACK_ROOK],
                 ]
                 if [starting_square_piece, ending_square_piece] in castling_pieces:
-                    return Stockfish.Capture.NO_CAPTURE
+                    return Capture.NO_CAPTURE
                 else:
-                    return Stockfish.Capture.DIRECT_CAPTURE
+                    return Capture.DIRECT_CAPTURE
         elif move_value[2:4] == self.get_fen_position().split()[
             3
         ] and starting_square_piece in [
-            Stockfish.Piece.WHITE_PAWN,
-            Stockfish.Piece.BLACK_PAWN,
+            Piece.WHITE_PAWN,
+            Piece.BLACK_PAWN,
         ]:
-            return Stockfish.Capture.EN_PASSANT
+            return Capture.EN_PASSANT
         else:
-            return Stockfish.Capture.NO_CAPTURE
+            return Capture.NO_CAPTURE
 
     def get_stockfish_full_version(self) -> float:
         """Returns Stockfish engine full version."""
@@ -1210,25 +1211,6 @@ class Stockfish:
     def __del__(self) -> None:
         Stockfish._del_counter += 1
         self.send_quit_command()
-
-    class Piece(Enum):
-        WHITE_PAWN = "P"
-        BLACK_PAWN = "p"
-        WHITE_KNIGHT = "N"
-        BLACK_KNIGHT = "n"
-        WHITE_BISHOP = "B"
-        BLACK_BISHOP = "b"
-        WHITE_ROOK = "R"
-        BLACK_ROOK = "r"
-        WHITE_QUEEN = "Q"
-        BLACK_QUEEN = "q"
-        WHITE_KING = "K"
-        BLACK_KING = "k"
-
-    class Capture(Enum):
-        DIRECT_CAPTURE = "direct capture"
-        EN_PASSANT = "en passant"
-        NO_CAPTURE = "no capture"
 
     @dataclass
     class BenchmarkParameters:
